modeling/ontology-agent/backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async HTTP
httpx>=0.26.0
aiohttp>=3.9.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# LLM
openai>=1.10.0

# Graph processing
networkx>=3.2.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0

# DNS/WHOIS
dnspython>=2.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
